sample/requirements.txt
Django==5.2.6
PyMySQL==1.1.1
PyPDF2==3.0.1
spacy==3.7.2
keybert==0.8.3
scikit-learn==1.3.2
openai==0.28.1
google-generativeai==0.3.2
google-cloud-aiplatform